17: add test script, refactor helper methods into transformers.py, and fix a breaking test related to dynamically created client datatable component
