+++ test
@@ -1,0 +1,2 @@
+#!/bin/bash
+pytest --cov-report term-missing  --cov=ujt -vv
