+++ tests/test_converters.py
@@ -432,17 +432,20 @@ def test_datatable_from_client():
             "id": "Status"
         },
     ]
-    expected_data = [{
-        "User Journey": "uj",
-        "Status": "HEALTHY",
-    }]
+    expected_data = [
+        {
+            "User Journey": "uj",
+            "Status": "HEALTHY",
+            "id": client.user_journeys[0].name,
+        }
+    ]
 
     table = ujt.converters.datatable_from_client(
         client,
         table_id=table_id,
     )
 
-    assert table.id == table_id  # pylint: disable=no-member
+    assert table.id == {"datatable-id": table_id}  # pylint: disable=no-member
     assert table.columns == expected_columns  # pylint: disable=no-member
     assert table.data == expected_data  # pylint: disable=no-member
     assert table.style_data_conditional == ujt.constants.DATATABLE_CONDITIONAL_STYLE  # pylint: disable=no-member
